Mark all classes as final
At least so far we don't implement any classes which are supposed to be
extended in production code.  To avoid additional BC implications wrt
code that would extend these classes we mark them as final.  Since we
still want to allow *some* of these classes to be faked for testing
purposes, we only use respective doc-block comments.