Start session before users may call Request::username()
The recently added `Request::username()` (#12) just looks for the user
name in `$_SESSION`.  However, this requires that the session had
already been started for the current request, which is not necessarily
the case.

We cater to that by probing for the session cookie, and start the
session if that is set, already in required_classes.php.  This still
does not solve the problem if someone calls `Request::username()` in an
earlier loaded required_classes.php, but misusing this file should only
be done in extraordinary cases, so we're likely good as is.